SQLite support for tg generate + dialect-neutral table codegen
Fixes surfaced along the way:
- Cardinality inference over-reported "one" for FK columns inside
  composite PKs/UNIQUEs (both dialects); SQLite rowid-alias FK
  columns were wrongly nullable ("maybe" instead of "one").
- PG stored-generated columns (`GENERATED ALWAYS AS ... STORED`)
  missed `generated: true`; ColumnInfo is now codegen-shaped
  (nullable/default/generated) instead of information_schema-ese.
- INSERT prunes columns no row provides so DB defaults apply
  natively; SQLite raises on heterogeneous rows (it has no per-row
  DEFAULT spelling) instead of silently inserting NULL.
- Examples now run codegen output as-is: SWC decorator transform in
  their vitest configs, tables regenerated with @expose().

+++ examples/sqlite/src/tables/teams.ts
@@ -1,11 +1,11 @@
 import { db } from "../db";
 import { expose } from "typegres";
-import { Int8, Text } from "typegres/postgres";
+import { Integer, Text } from "typegres/sqlite";
 import { Dogs } from "./dogs";
 
 export class Teams extends db.Table("teams") {
   // @generated-start
-  @expose() id = (Int8<1>).column({ nonNull: true, generated: true });
+  @expose() id = (Integer<1>).column({ nonNull: true, generated: true });
   @expose() name = (Text<1>).column({ nonNull: true });
   // relations
   @expose() dogs() { return Dogs.scope(Teams.contextOf(this)).where(({ dogs }) => dogs.team_id.eq(this.id)).cardinality("many"); }
